refactor: extract isBondBetweenMonomers IIFE into utils
Addresses review feedback on PR #11373 to avoid a long inline IIFE
in the component body.
